finished api

+++ risk-calc-app/Controllers/StockController.cs
@@ -2,7 +2,9 @@
 using risk_calc_app.Data;
 using risk_calc_app.DTOs.Stock;
 using risk_calc_app.Interfaces;
+using risk_calc_app.Mappers;
 using risk_calc_app.Models;
+using risk_calc_app.DTOs;
 
 namespace risk_calc_app.Controllers
 {
@@ -12,51 +14,76 @@ namespace risk_calc_app.Controllers
     {
         private readonly RiskCalcAppDbContext _context;
         private readonly IStockRepository _stockRepo;
+        private readonly IPortfolioRepository _portfolioRepo;
 
-        public StockController(RiskCalcAppDbContext context, IStockRepository stockRepo)
+        public StockController(RiskCalcAppDbContext context, IStockRepository stockRepo, IPortfolioRepository postfolioRepo)
         {
             _stockRepo = stockRepo;
+            _portfolioRepo = postfolioRepo;
             _context = context;
         }
 
         //GET api/stocks
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<StockItem>>> Get()
+        public async Task<ActionResult<List<StockDto>>> Get()
         {
             var stocks = await _stockRepo.GetAllStocksAsync();
-            return Ok(stocks);
+            var stockDtos = stocks.Select(i => i.ToStockDto()).ToList();
+            return Ok(stockDtos);
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<StockItem>> GetById(int id)
+        public async Task<ActionResult<StockDto>> GetById(int id)
         {
             var stockItem = await _stockRepo.GetStockByIdAsync(id);
-            return Ok(stockItem);
+
+            if (stockItem == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(stockItem.ToStockDto());
         }
 
         //POST api/stocks
-        [HttpPost]
-        public async Task<ActionResult<StockItem>> Create(StockDto stockDto)
+        [HttpPost("{portfolioId}")]
+        public async Task<ActionResult<StockItem>> Create([FromRoute] int portfolioId, [FromBody] CreateStockDto createStockDto)
         {
-            await _stockRepo.CreateStockAsync(stockDto);
+            if(!await _portfolioRepo.PortfolioExists(portfolioId))
+            {
+                return BadRequest("Stock does not exist.");
+            }
 
-            return Ok(stockDto);
+            var stock = createStockDto.ToStockFromCreateStockDto(portfolioId);
+
+            await _stockRepo.CreateStockAsync(stock);
+            return CreatedAtAction(nameof(GetById), new {id =  stock.Id}, stock.ToStockDto());
         }
 
         //PUT api/stocks
         [HttpPut("{id}")]
-        public async Task<ActionResult> Put(int id, [FromBody] StockDto stockDto)
+        public async Task<ActionResult> Put([FromRoute] int id, [FromBody] UpdateStockDto updateStockDto)
         {
-            await _stockRepo.UpdateStockByIdAsync(id, stockDto);
-            return NoContent();
+            if(!await _stockRepo.StockExists(id))
+            {
+                return NotFound();
+            }
+            var stock = updateStockDto.ToStockFromUpdateStockDto();
+            await _stockRepo.UpdateStockByIdAsync(id, stock);
+            return Ok(stock.ToStockDto());
         }
 
         //DELETE api/stocks
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(int id)
         {
-            await _stockRepo.DeleteStockByIdAsync(id);
-            return NoContent();
+            if (!await _stockRepo.StockExists(id))
+            {
+                return NotFound();
+            }
+
+            var stock = await _stockRepo.DeleteStockByIdAsync(id);
+            return Ok(stock);
         }
     }
 }
